requirements.txt
Django==5.1.11
djangorestframework==3.15.2
Pillow==11.3.0
django-cors-headers==4.7.0
python-decouple==3.8
requests==2.32.0
lxml==5.3.0
reportlab==4.2.5
python-magic==0.4.27 